+++ src/MyLab.Log.Dsl/LoggerExtensions.cs
@@ -11,7 +11,7 @@ namespace MyLab.Log.Dsl
         /// <summary>
         /// Create DSL-logger for core logger
         /// </summary>
-        [Obsolete("Use DI instead: services.AddDslLog() and DI with IDslLogger<TCat>", true)]
+        [Obsolete("Use DI instead: services.AddLogging(l => l.AddDslLog()) and DI with IDslLogger<TCat>", true)]
         public static IDslLogger Dsl(this ILogger coreLogger)
         {
             throw new NotSupportedException();